Updated IDEIndexCompletionItem for Xcode 7.1.

+++ KSImageNamed/KSImageNamedIndexCompletionItem.m
@@ -43,7 +43,7 @@
     
 }
 
-- (long long)priority
+- (double)priority
 {
     return 9999;
 }
